feat(autocount): GRN ingest - Decimal qty + supplier + dry_run/verdict/idempotency (S17)
Bring the last legacy AutoCount ingest (POST /api/v1/external/grn/) into the
adopt-by-source_ref verdict pattern the other 12 entities already use.

1a Decimal widen (migration 310, chains onto 309): picking_lines
   {qty_accepted,qty_rejected,quantity_expected,quantity_picked} and
   picking_headers {total_items_picked,total_items_discrepancy} -> numeric(15,4)
   so AutoCount fractional quantities ('2.5') stop truncating. Drops+recreates
   the generated quantity_discrepancy AND the dependent scm.receipt_lead_v view
   around the ALTER (Postgres cannot alter a column a view reads). BE schema and
   FE grn-schema/types swept int -> Decimal.

1b Supplier: picking_headers gains supplier_code + supplier_id (FK->suppliers,
   ON DELETE SET NULL). Resolve-or-null - a supplier miss keeps the code with
   supplier_id NULL, never a 400. GRN detail shows a Supplier row.

1c Endpoint via new GrnIngestService (clone of the DO ingest, entity_type
   'picking_headers'): source_ref present -> 200-always verdict envelope,
   idempotent update-in-place, dry_run rolls back (service savepoint + endpoint
   top-level), product/warehouse miss -> retryable, uom best-effort. source_ref
   absent -> legacy create-only 201 (back-compat); legacy + dry_run -> 400.

Raw picking_lines insert carries the header company_id (ORM auto-stamp is
bypassed by raw SQL) and sets synced_to_excel explicitly (no server default).

Tests: tests/test_ingest_grn.py (12 cases) + FE grn-schema.test.ts green;
72 procurement-file pytest green; live HTTP + Playwright verified.

+++ sorento_crm_frontend/app/(protected)/procurement-management/grn/forms/grn-schema.ts
@@ -2,8 +2,8 @@ import { z } from 'zod';
 
 export const pickingLineSchema = z.object({
   product_id: z.string().min(1, 'Product is required'),
-  quantity_expected: z.coerce.number().int().min(0, 'Quantity expected must be 0 or more'),
-  quantity_picked: z.coerce.number().int().min(0, 'Quantity received must be 0 or more'),
+  quantity_expected: z.coerce.number().min(0, 'Quantity expected must be 0 or more'),
+  quantity_picked: z.coerce.number().min(0, 'Quantity received must be 0 or more'),
   source_warehouse_id: z.string().min(1, 'Location is required'),
   spo_allocation_id: z.string().optional(),
   picked_condition: z.string().optional(),
